Sort Search Read Write actions are performed(specified for also Turkish)
It reads the content of a user-specified text file and processes the words, excluding those listed in an "ignore" file. It calculates and displays the occurrences of unique words in the file, ignoring words from the "ignore" list. The program writes the unique words and their frequencies to a new file named "words.txt." Additionally, it sorts and displays the unique words in alphabetical order. Users can enter a word to check its existence using a recursive search function. It is specified for Turkish.